data/archive/requirements 2.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Kafka & Messaging
kafka-python==2.0.2
aiokafka==0.10.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# ML/AI
openai==1.7.2
anthropic==0.8.1
langchain==0.1.0
langchain-openai==0.0.2
tiktoken==0.5.2

# Data Processing
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# CLI
click==8.1.7
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.39.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2
arrow==1.3.0
python-dateutil==2.8.2

# Development
ipython==8.20.0
jupyter==1.0.0
